feat: Update authentication endpoints
- Increase access token expiration to 3 days
- Add user object to login response
- Remove duplicate /auth/token endpoint
- Move /auth/refresh endpoint to AuthController
